+++ flutter_navigation_generator/lib/src/code_builder/import_builder.dart
@@ -15,17 +15,23 @@ class ImportBuilder {
   });
 
   Iterable<Directive> generate() {
-    final imports = <String?>{'package:flutter/material.dart', 'dart:convert'};
+    final imports = <String?>{
+      'package:flutter/material.dart',
+      'dart:convert',
+      if (routes.any((route) => route.guards.isNotEmpty))
+        'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart',
+    };
     imports.add(typeRefer(pageType, targetFile: targetFile).url);
-    imports.addAll(routes
-        .map((route) => [
-              typeRefer(route.routeWidget, targetFile: targetFile).url,
-              typeRefer(route.pageType, targetFile: targetFile).url,
-              typeRefer(route.returnType, targetFile: targetFile).url,
-              ...route.parameters
-                  .map((e) => typeRefer(e, targetFile: targetFile).url),
-            ])
-        .expand((element) => element));
+    imports.addAll(routes.expand(
+      (route) => [
+        typeRefer(route.routeWidget, targetFile: targetFile).url,
+        typeRefer(route.pageType, targetFile: targetFile).url,
+        typeRefer(route.returnType, targetFile: targetFile).url,
+        ...route.parameters
+            .map((e) => typeRefer(e, targetFile: targetFile).url),
+        ...route.guards.map((e) => typeRefer(e, targetFile: targetFile).url),
+      ],
+    ));
     return imports.whereType<String>().map(Directive.import);
   }
 }