+++ flutter_navigation_generator/lib/src/code_builder/on_generate_route_builder.dart
@@ -2,6 +2,7 @@ import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
+import 'package:flutter_navigation_generator/src/utils/case_utils.dart';
 import 'package:flutter_navigation_generator/src/utils/route_config_extension.dart';
 import 'package:flutter_navigation_generator/src/utils/utils.dart';
 import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
@@ -29,26 +30,50 @@ class OnGenerateRouteBuilder {
   }
 
   String _generateRoute(RouteConfig route) {
-    final constructor = route.constructorName == route.routeWidget.className || route.constructorName.isEmpty
+    final constructor = route.constructorName == route.routeWidget.className ||
+            route.constructorName.isEmpty
         ? route.routeWidget.className
         : '${route.routeWidget.className}.${route.constructorName}';
-    final constructorCall = '$constructor(${route.parameters.asMap().map((_, p) {
-          final nullableSuffix = p.isNullable ? '?' : '';
-          final convertFromString = _convertFromString(p, 'arguments[\'${p.argumentName}\']');
-          return MapEntry(
-            p.argumentName,
-            (p.className == 'Key' || p.className == 'String')
-                ? "arguments['${p.argumentName}'] as ${typeRefer(p).symbol}$nullableSuffix"
-                : "arguments['${p.argumentName}'] is String ? $convertFromString : arguments['${p.argumentName}'] as ${typeRefer(p).symbol}$nullableSuffix",
-          );
-        }).entries.map((e) => '${e.key}: ${e.value},').join('')})';
-    return 'return ${_withPageType(route, constructorCall)};';
+    final constructorCall =
+        '$constructor(${route.parameters.asMap().map((_, p) {
+              final nullableSuffix = p.isNullable ? '?' : '';
+              final convertFromString =
+                  _convertFromString(p, 'arguments[\'${p.argumentName}\']');
+              return MapEntry(
+                p.argumentName,
+                (p.className == 'Key' || p.className == 'String')
+                    ? "arguments['${p.argumentName}'] as ${typeRefer(p).symbol}$nullableSuffix"
+                    : "arguments['${p.argumentName}'] is String ? $convertFromString : arguments['${p.argumentName}'] as ${typeRefer(p).symbol}$nullableSuffix",
+              );
+            }).entries.map((e) => '${e.key}: ${e.value},').join('')})';
+
+    var guardsCode = '';
+    if (route.guards.isNotEmpty) {
+      for (final guard in route.guards) {
+        final insanceName = CaseUtil(guard.className).camelCase;
+        guardsCode +=
+            'final $insanceName = guards.whereType<${typeRefer(guard).symbol}>().first;\n';
+        guardsCode += 'if (!$insanceName.value) {\n';
+        guardsCode += '  return onGenerateRoute(RouteSettings(\n';
+        guardsCode += '    arguments: settings.arguments,\n';
+        guardsCode += '    name: $insanceName.alternativeRoute,\n';
+        guardsCode += '  ));\n';
+        guardsCode += '}\n';
+      }
+    }
+    return '${guardsCode}return ${_withPageType(route, constructorCall)};';
   }
 
   Method generate() {
-    final pageRoutes = routes.where((r) => r.generatePageRoute && r.navigationType != NavigationType.bottomSheet && r.navigationType != NavigationType.dialog).toList();
+    final pageRoutes = routes
+        .where((r) =>
+            r.generatePageRoute &&
+            r.navigationType != NavigationType.bottomSheet &&
+            r.navigationType != NavigationType.dialog)
+        .toList();
     for (final pageRoute in pageRoutes.toList()) {
-      final pageRoute2 = pageRoutes.firstWhere((element) => element.routeName == pageRoute.routeName);
+      final pageRoute2 = pageRoutes
+          .firstWhere((element) => element.routeName == pageRoute.routeName);
       if (pageRoute != pageRoute2) {
         pageRoutes.remove(pageRoute);
       }
@@ -66,7 +91,8 @@ class OnGenerateRouteBuilder {
         )
         ..body = Block.of([
           if (pageRoutes.isNotEmpty) ...[
-            const Code('''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
+            const Code(
+                '''final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};
     final settingsUri = Uri.parse(settings.name ?? '');
     settingsUri.queryParameters.forEach((key, value) {
       arguments[key] ??= value;
@@ -74,28 +100,38 @@ class OnGenerateRouteBuilder {
             Code(
                 'switch (settingsUri.path) {${pageRoutes.where((route) => !route.routeNameContainsParameters).map((route) => 'case RouteNames.${route.asRouteName}: ${_generateRoute(route)}').join('')}}'),
           ],
-          if (pageRoutes.any((element) => element.routeNameContainsParameters)) ...[
+          if (pageRoutes
+              .any((element) => element.routeNameContainsParameters)) ...[
             const Code('final pathSegments = settingsUri.pathSegments;'),
             ...pageRoutes
                 .where((pageRoute) => pageRoute.routeNameContainsParameters)
-                .groupListsBy((pageRoute) => pageRoute.routeName.pathSegments.length)
+                .groupListsBy(
+                    (pageRoute) => pageRoute.routeName.pathSegments.length)
                 .entries
                 .sorted((a, b) => -a.key.compareTo(b.key))
                 .map((group) {
               final pathSegments = group.key;
               var code = 'if (pathSegments.length == $pathSegments) {';
-              final pageRoutesMap = group.value.asMap().map((key, value) => MapEntry(value, value.routeName.parametersFromRouteName.length));
-              final pageRoutes = pageRoutesMap.entries.sorted((a, b) => a.value.compareTo(b.value)).map((e) => e.key).toList();
+              final pageRoutesMap = group.value.asMap().map((key, value) =>
+                  MapEntry(
+                      value, value.routeName.parametersFromRouteName.length));
+              final pageRoutes = pageRoutesMap.entries
+                  .sorted((a, b) => a.value.compareTo(b.value))
+                  .map((e) => e.key)
+                  .toList();
               for (final pageRoute in pageRoutes) {
                 final pathSegments = pageRoute.routeName.pathSegments;
-                final hasRigidSegments = pathSegments.any((element) => !element.startsWith(':'));
+                final hasRigidSegments =
+                    pathSegments.any((element) => !element.startsWith(':'));
                 if (hasRigidSegments) {
                   code += 'if (';
                   code += pathSegments
                       .asMap()
                       .entries
-                      .where((pathSegment) => !pathSegment.value.startsWith(':'))
-                      .map((pathSegment) => 'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
+                      .where(
+                          (pathSegment) => !pathSegment.value.startsWith(':'))
+                      .map((pathSegment) =>
+                          'pathSegments[${pathSegment.key}] == \'${pathSegment.value}\'')
                       .join(' && ');
                   code += ') {';
                 }
@@ -103,7 +139,8 @@ class OnGenerateRouteBuilder {
                     .asMap()
                     .entries
                     .where((pathSegment) => pathSegment.value.startsWith(':'))
-                    .map((pathSegment) => 'arguments[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
+                    .map((pathSegment) =>
+                        'arguments[\'${pathSegments[pathSegment.key].substring(1)}\'] = pathSegments[${pathSegment.key}];')
                     .join('\n');
                 code += _generateRoute(pageRoute);
                 if (hasRigidSegments) code += '}';
@@ -113,7 +150,8 @@ class OnGenerateRouteBuilder {
             }),
           ],
           if (unknownRoute != null) ...[
-            Code('return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
+            Code(
+                'return ${_withPageType(null, '${typeRefer(unknownRoute!).symbol!}()')};'),
           ] else ...[
             const Code('return null;'),
           ],