+++ flutter_navigation_generator/lib/src/code_builder/route_builder.dart
@@ -54,7 +54,8 @@ class RouteBuilder {
     if (route.methodNameIsDefinedByAnnotation) {
       name = route.methodName;
     } else {
-      name = '$namePrefix${CaseUtil(route.methodName.replaceAll(RegExp(':[a-zA-Z0-9]*'), '')).upperCamelCase}';
+      name =
+          '$namePrefix${CaseUtil(route.methodName.replaceAll(RegExp(':[a-zA-Z0-9]*'), '')).upperCamelCase}';
     }
     return Method(
       (b) => b
@@ -68,7 +69,9 @@ class RouteBuilder {
                 ..name = p.argumentName
                 ..named = true
                 ..required = p.isRequired && route.defaultValues[p.name] == null
-                ..defaultTo = route.defaultValues[p.name] != null ? Reference(route.defaultValues[p.name]! as String).code : null
+                ..defaultTo = route.defaultValues[p.name] != null
+                    ? Reference(route.defaultValues[p.name]! as String).code
+                    : null
                 ..type = typeRefer(p),
             ),
           ),
@@ -114,35 +117,48 @@ class RouteBuilder {
           path = Reference('RouteNames.${route.asRouteName}').call(
               [],
               (parameters.asMap().map((_, parameter) {
-                final argument = route.parameters.firstWhereOrNull((element) => element.name == parameter);
+                final argument = route.parameters
+                    .firstWhereOrNull((element) => element.name == parameter);
                 if (argument == null) return MapEntry(parameter, null);
-                return MapEntry(parameter, Reference(_generateParameterValue(argument)));
+                return MapEntry(
+                    parameter, Reference(_generateParameterValue(argument)));
               })
                     ..removeWhere((key, value) => value == null))
                   .cast());
         } else {
           path = Reference('RouteNames.${route.asRouteName}');
         }
-        final queryParameters =
-            route.parameters.where((element) => element.className != 'Key' && !parameters.contains(element.argumentName)).toList().asMap().map((_, p) => MapEntry(
+        final queryParameters = route.parameters
+            .where((element) =>
+                element.className != 'Key' &&
+                !parameters.contains(element.argumentName))
+            .toList()
+            .asMap()
+            .map((_, p) => MapEntry(
                   "'${p.argumentName}'",
                   _generateParameterValue(p),
                 ));
 
-        final queryParametersRemoveNull = route.parameters.any((element) => element.className != 'Key' && !parameters.contains(element.argumentName) && element.isNullable)
+        final queryParametersRemoveNull = route.parameters.any((element) =>
+                element.className != 'Key' &&
+                !parameters.contains(element.argumentName) &&
+                element.isNullable)
             ? '..removeWhere((_, v) => v == null)'
             : '';
 
         final bodyCall = TypeReference(
           (b) => b
-            ..symbol = 'navigatorKey.currentState?.${route.navigationType.navigatorMethod}'
+            ..symbol =
+                'navigatorKey.currentState?.${route.navigationType.navigatorMethod}'
             ..types.addAll(
               [
                 const Reference('dynamic'),
                 if (route.navigationType == NavigationType.pushReplacement ||
                     route.navigationType == NavigationType.popAndPush ||
-                    route.navigationType == NavigationType.restorablePushReplacement ||
-                    route.navigationType == NavigationType.restorablePopAndPush) ...[
+                    route.navigationType ==
+                        NavigationType.restorablePushReplacement ||
+                    route.navigationType ==
+                        NavigationType.restorablePopAndPush) ...[
                   const Reference('dynamic'),
                 ],
               ],
@@ -158,22 +174,32 @@ class RouteBuilder {
                     [],
                     {
                       'path': path,
-                      'queryParameters': Reference('$queryParameters$queryParametersRemoveNull'),
+                      'queryParameters': Reference(
+                          '$queryParameters$queryParametersRemoveNull'),
                     },
                   ).property('toString()'),
-            if (route.navigationType == NavigationType.pushAndReplaceAll || route.navigationType == NavigationType.restorablePushAndReplaceAll) ...[
+            if (route.navigationType == NavigationType.pushAndReplaceAll ||
+                route.navigationType ==
+                    NavigationType.restorablePushAndReplaceAll) ...[
               const Reference('(_) => false'),
             ],
           ],
           {
-            'arguments': Reference('${route.parameters.asMap().map((_, p) => MapEntry("'${p.argumentName}'", p.argumentName))}'),
+            'arguments': Reference(
+                '${route.parameters.asMap().map((_, p) => MapEntry("'${p.argumentName}'", p.argumentName))}'),
           },
         );
         Code body;
         if (route.returnType != null) {
           body = Block((b) => b
-            ..statements.add(declareFinal('result', type: const Reference('dynamic')).assign(bodyCall.awaited).statement)
-            ..statements.add(const Reference('result').asA(typeRefer(route.returnType, forceNullable: true)).returned.statement));
+            ..statements.add(
+                declareFinal('result', type: const Reference('dynamic'))
+                    .assign(bodyCall.awaited)
+                    .statement)
+            ..statements.add(const Reference('result')
+                .asA(typeRefer(route.returnType, forceNullable: true))
+                .returned
+                .statement));
         } else {
           body = bodyCall.code;
         }
@@ -199,16 +225,22 @@ class RouteBuilder {
       body: TypeReference(
         (b) => b
           ..symbol = bodyCall
-          ..types.add(route.returnType == null ? const Reference('dynamic') : typeRefer(route.returnType!)),
+          ..types.add(route.returnType == null
+              ? const Reference('dynamic')
+              : typeRefer(route.returnType!)),
       ).call(
         [],
         {
           'widget': Reference(
-            route.constructorName == route.routeWidget.className || route.constructorName.isEmpty
+            route.constructorName == route.routeWidget.className ||
+                    route.constructorName.isEmpty
                 ? route.routeWidget.className
                 : '${route.routeWidget.className}.${route.constructorName}',
             typeRefer(route.routeWidget).url,
-          ).call([], route.parameters.asMap().map((_, p) => MapEntry(p.argumentName, Reference(p.argumentName)))),
+          ).call(
+              [],
+              route.parameters.asMap().map((_, p) =>
+                  MapEntry(p.argumentName, Reference(p.argumentName)))),
         },
       ).code,
     );
@@ -236,7 +268,8 @@ class RouteBuilder {
             ..name = 'goBack'
             ..lambda = true
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.pop').call([]).code,
+            ..body =
+                const Reference('navigatorKey.currentState?.pop').call([]).code,
         ),
         Method(
           (b) => b
@@ -250,7 +283,8 @@ class RouteBuilder {
                 ..type = const Reference('T?'),
             ))
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.pop').call([const Reference('result')]).code,
+            ..body = const Reference('navigatorKey.currentState?.pop')
+                .call([const Reference('result')]).code,
         ),
         Method(
           (b) => b
@@ -264,7 +298,8 @@ class RouteBuilder {
                   ..requiredParameters.add(const Reference('Route<dynamic>'))),
             ))
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.popUntil').call([const Reference('predicate')]).code,
+            ..body = const Reference('navigatorKey.currentState?.popUntil')
+                .call([const Reference('predicate')]).code,
         ),
         Method(
           (b) => b
@@ -293,7 +328,8 @@ class RouteBuilder {
             ..returns = const Reference('Future<T?>')
             ..body = const Reference('showDialog<T>').call([], {
               'context': const Reference('navigatorKey.currentContext!'),
-              'builder': const Reference('(_) => widget ?? const SizedBox.shrink()'),
+              'builder':
+                  const Reference('(_) => widget ?? const SizedBox.shrink()'),
             }).code,
         ),
         Method(
@@ -311,7 +347,8 @@ class RouteBuilder {
             ..returns = const Reference('Future<T?>')
             ..body = const Reference('showModalBottomSheet<T>').call([], {
               'context': const Reference('navigatorKey.currentContext!'),
-              'builder': const Reference('(_) => widget ?? const SizedBox.shrink()'),
+              'builder':
+                  const Reference('(_) => widget ?? const SizedBox.shrink()'),
             }).code,
         ),
       ];