+++ flutter_navigation_generator/lib/src/utils/case_utils.dart
@@ -5,13 +5,16 @@ class CaseUtil {
   late String originalText;
   late List<String> _words;
 
-  CaseUtil(String text, {String? alternativeText, List<String> removeSuffixes = const []}) {
+  CaseUtil(String text,
+      {String? alternativeText, List<String> removeSuffixes = const []}) {
     for (final removeSuffix in removeSuffixes) {
       if (text.toLowerCase().endsWith(removeSuffix.toLowerCase())) {
         text = text.substring(0, text.length - removeSuffix.length);
       }
-      if (alternativeText != null && alternativeText.toLowerCase().endsWith(removeSuffix.toLowerCase())) {
-        alternativeText = alternativeText.substring(0, alternativeText.length - removeSuffix.length);
+      if (alternativeText != null &&
+          alternativeText.toLowerCase().endsWith(removeSuffix.toLowerCase())) {
+        alternativeText = alternativeText.substring(
+            0, alternativeText.length - removeSuffix.length);
       }
     }
     originalText = text;
@@ -29,7 +32,9 @@ class CaseUtil {
 
     for (var i = 0; i < text.length; i++) {
       final char = String.fromCharCode(text.codeUnitAt(i));
-      final nextChar = i + 1 == text.length ? null : String.fromCharCode(text.codeUnitAt(i + 1));
+      final nextChar = i + 1 == text.length
+          ? null
+          : String.fromCharCode(text.codeUnitAt(i + 1));
 
       if (_symbolRegex.hasMatch(char)) {
         continue;
@@ -37,7 +42,9 @@ class CaseUtil {
 
       sb.write(char);
 
-      final isEndOfWord = nextChar == null || (_upperAlphaRegex.hasMatch(nextChar) && !isAllCaps) || _symbolRegex.hasMatch(nextChar);
+      final isEndOfWord = nextChar == null ||
+          (_upperAlphaRegex.hasMatch(nextChar) && !isAllCaps) ||
+          _symbolRegex.hasMatch(nextChar);
 
       if (isEndOfWord) {
         words.add(sb.toString());
@@ -58,7 +65,8 @@ class CaseUtil {
 
   String get upperCamelCase => _uppserCamelCase();
 
-  String _getSnakeCase({String separator = '_'}) => _words.map((word) => word.toLowerCase()).toList().join(separator);
+  String _getSnakeCase({String separator = '_'}) =>
+      _words.map((word) => word.toLowerCase()).toList().join(separator);
 
   String _getCamelCase({String separator = ''}) {
     final words = _words.map(_upperCaseFirstLetter).toList();